Fix NSQ sleep/requeue functionality
